+++ Cql/CodeGeneration.NET/GlobalUsings.cs
@@ -19,3 +19,9 @@ global using Microsoft.CodeAnalysis.Text;
 global using BinaryExpression = System.Linq.Expressions.BinaryExpression;
 global using Expression = System.Linq.Expressions.Expression;
 global using UnaryExpression = System.Linq.Expressions.UnaryExpression;
+
+global using ElmLibrary = Hl7.Cql.Elm.Library;
+
+global using ElmToolkitConversionDictionary = System.Collections.Immutable.ImmutableDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, Hl7.Cql.CodeGeneration.NET.Toolkit.ElmToolkitConversionRecord>;
+global using ElmToolkitConversionReadOnlyDictionary = System.Collections.Generic.IReadOnlyDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, Hl7.Cql.CodeGeneration.NET.Toolkit.ElmToolkitConversionRecord>;
+
