+++ Cql/CodeGeneration.NET/Toolkit/Extensions/ElmToolkitExtensions.Adding.cs
@@ -6,21 +6,34 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
  */
 
-using CqlSdkPrototype.Cql;
-using CqlSdkPrototype.Cql.Extensions;
+using Hl7.Cql.CqlToElm.Toolkit;
+using Hl7.Cql.CqlToElm.Toolkit.Extensions;
 using Hl7.Cql.Runtime;
+using Microsoft.Extensions.Logging;
 
-#pragma warning disable RS0027
-
-namespace CqlSdkPrototype.Elm.Extensions;
+namespace Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 
 public static partial class ElmToolkitExtensions
 {
+    /// <summary>
+    /// Adds ELM libraries from a CQL toolkit to the specified ELM toolkit.
+    /// </summary>
+    /// <param name="elmToolkit">The ELM toolkit to add libraries to.</param>
+    /// <param name="cqlToolkit">The CQL toolkit containing the libraries to add.</param>
+    /// <returns>The updated ELM toolkit.</returns>
     public static ElmToolkit AddElmFromCqlToolkit(
         this ElmToolkit elmToolkit,
         CqlToolkit cqlToolkit) =>
         elmToolkit.AddElmLibraries(cqlToolkit.GetCqlToolkitResults().Select(t => t.ElmLibrary));
 
+    /// <summary>
+    /// Adds an ELM file from a specified directory to the ELM toolkit.
+    /// </summary>
+    /// <param name="elmToolkit">The ELM toolkit to add the file to.</param>
+    /// <param name="directory">The directory containing the ELM file.</param>
+    /// <param name="versionedLibraryIdentifier">The identifier of the library file.</param>
+    /// <returns>The updated ELM toolkit.</returns>
+    /// <exception cref="FileNotFoundException">Thrown when the specified file is not found.</exception>
     public static ElmToolkit AddElmFileInDirectory(
         this ElmToolkit elmToolkit,
         DirectoryInfo directory,
@@ -41,6 +54,12 @@ public static partial class ElmToolkitExtensions
         return elmToolkit.AddElmFile(file);
     }
 
+    /// <summary>
+    /// Adds multiple ELM files to the ELM toolkit.
+    /// </summary>
+    /// <param name="elmToolkit">The ELM toolkit to add the files to.</param>
+    /// <param name="files">The collection of ELM files to add.</param>
+    /// <returns>The updated ELM toolkit.</returns>
     public static ElmToolkit AddElmFiles(
         this ElmToolkit elmToolkit,
         IEnumerable<FileInfo> files)
@@ -56,6 +75,14 @@ public static partial class ElmToolkitExtensions
         return elmToolkit.AddElmLibraries(libraries);
     }
 
+    /// <summary>
+    /// Adds ELM files from a specified directory to the ELM toolkit.
+    /// </summary>
+    /// <param name="elmToolkit">The ELM toolkit to add the files to.</param>
+    /// <param name="directory">The directory containing the ELM files.</param>
+    /// <param name="options">Optional enumeration options for file retrieval.</param>
+    /// <param name="filePredicate">Optional predicate to filter files.</param>
+    /// <returns>The updated ELM toolkit.</returns>
     public static ElmToolkit AddElmFilesFromDirectory(
         this ElmToolkit elmToolkit,
         DirectoryInfo directory,
@@ -67,12 +94,29 @@ public static partial class ElmToolkitExtensions
         return elmToolkit.AddElmFiles(files);
     }
 
+    /// <summary>
+    /// Adds an ELM file with its dependencies to the ELM toolkit.
+    /// </summary>
+    /// <param name="elmToolkit">The ELM toolkit to add the file to.</param>
+    /// <param name="file">The ELM file to add.</param>
+    /// <param name="options">Optional enumeration options for dependency retrieval.</param>
+    /// <returns>The updated ELM toolkit.</returns>
+    /// <exception cref="NotImplementedException">Thrown when the method is not implemented.</exception>
     private static ElmToolkit AddElmFileWithDependencies(
         this ElmToolkit elmToolkit,
         FileInfo file,
         EnumerationOptions? options) =>
         throw new NotImplementedException();
 
+    /// <summary>
+    /// Adds an ELM file with its dependencies from a specified directory to the ELM toolkit.
+    /// </summary>
+    /// <param name="elmToolkit">The ELM toolkit to add the file to.</param>
+    /// <param name="directory">The directory containing the ELM file and its dependencies.</param>
+    /// <param name="fileName">The identifier of the library file.</param>
+    /// <param name="options">Optional enumeration options for dependency retrieval.</param>
+    /// <returns>The updated ELM toolkit.</returns>
+    /// <exception cref="NotImplementedException">Thrown when the method is not implemented.</exception>
     private static ElmToolkit AddElmFileWithDependencies(
         this ElmToolkit elmToolkit,
         DirectoryInfo directory,
@@ -80,6 +124,12 @@ public static partial class ElmToolkitExtensions
         EnumerationOptions? options) =>
         throw new NotImplementedException();
 
+    /// <summary>
+    /// Adds a single ELM file to the ELM toolkit.
+    /// </summary>
+    /// <param name="elmToolkit">The ELM toolkit to add the file to.</param>
+    /// <param name="file">The ELM file to add.</param>
+    /// <returns>The updated ELM toolkit.</returns>
     public static ElmToolkit AddElmFile(
         this ElmToolkit elmToolkit,
         FileInfo file) =>
