+++ Cql/CodeGeneration.NET/Toolkit/Extensions/ElmToolkitExtensions.cs
@@ -7,13 +7,21 @@
  */
 
 using Hl7.Cql.Abstractions.Infrastructure;
-using Hl7.Cql.CodeGeneration.NET;
 using Hl7.Cql.Runtime;
 
-namespace CqlSdkPrototype.Elm.Extensions;
+namespace Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 
+/// <summary>
+/// Provides extension methods for the <see cref="ElmToolkit"/> class.
+/// </summary>
 public static partial class ElmToolkitExtensions
 {
+    /// <summary>
+    /// Configures the <see cref="ElmToolkit"/> to ignore exceptions.
+    /// </summary>
+    /// <param name="cqlToolkit">The <see cref="ElmToolkit"/> instance.</param>
+    /// <param name="stopAfterFirstException">If set to <c>true</c>, stops after the first exception; otherwise, continues.</param>
+    /// <returns>The configured <see cref="ElmToolkit"/> instance.</returns>
     public static ElmToolkit SetExceptionHandlingToIgnore(this ElmToolkit cqlToolkit, bool stopAfterFirstException = false) =>
         cqlToolkit.Reconfigure(o => o with
         {
@@ -22,31 +30,44 @@ public static partial class ElmToolkitExtensions
                                                     : ErroredEnumerationContinuation.Continue
         });
 
+    /// <summary>
+    /// Configures the <see cref="ElmToolkit"/> to use embedded debug information.
+    /// </summary>
+    /// <param name="cqlToolkit">The <see cref="ElmToolkit"/> instance.</param>
+    /// <returns>The configured <see cref="ElmToolkit"/> instance.</returns>
     public static ElmToolkit SetAssemblyDebugInformationToEmbedded(this ElmToolkit cqlToolkit) =>
         cqlToolkit.Reconfigure(o => o with
         {
             AssemblyCompilerDebugInformationFormat = AssemblyCompilerDebugInformationFormat.Embedded
         });
 
+    /// <summary>
+    /// Configures the <see cref="ElmToolkit"/> to use portable PDB debug information.
+    /// </summary>
+    /// <param name="cqlToolkit">The <see cref="ElmToolkit"/> instance.</param>
+    /// <returns>The configured <see cref="ElmToolkit"/> instance.</returns>
     public static ElmToolkit SetAssemblyDebugInformationToPortablePdb(this ElmToolkit cqlToolkit) =>
         cqlToolkit.Reconfigure(o => o with
         {
             AssemblyCompilerDebugInformationFormat = AssemblyCompilerDebugInformationFormat.PortablePdb
         });
 
-
+    /// <summary>
+    /// Gets the results of ELM to assembly conversions.
+    /// </summary>
+    /// <param name="elmToolkit">The <see cref="ElmToolkit"/> instance.</param>
+    /// <returns>An enumerable of <see cref="ElmToolkitResultRecord"/> containing the conversion results.</returns>
     public static IEnumerable<ElmToolkitResultRecord> GetElmToAssemblyResults(
         this ElmToolkit elmToolkit) =>
         elmToolkit.Conversions.Values
                   .SelectWhere(t => t switch
                   {
                       {
-                          LibraryIdentifier: {} libId,
-                          ResultCSharpSourceCode: {} csharp,
-                          ResultAssemblyBinary: {} asm,
+                          LibraryIdentifier: { } libId,
+                          ResultCSharpSourceCode: { } csharp,
+                          ResultAssemblyBinary: { } asm,
                           ResultDebugSymbolsBinary: var dbg
                       } => (true, new ElmToolkitResultRecord(libId, csharp, asm, dbg)),
                       _ => default,
                   });
-
 }
