+++ Cql/CodeGeneration.NET/Toolkit/Internal/ElmToolkitServices.cs
@@ -7,15 +7,17 @@
  */
 
 using Hl7.Cql.Abstractions;
-using Hl7.Cql.CodeGeneration.NET;
 using Hl7.Cql.Compiler;
 using Hl7.Cql.Conversion;
 using Hl7.Cql.Fhir;
 using Hl7.Cql.Runtime.Hosting;
 using Hl7.Cql.Runtime.Logging;
 using Hl7.Fhir.Introspection;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Logging;
 
-namespace CqlSdkPrototype.Elm.Internal;
+namespace Hl7.Cql.CodeGeneration.NET.Toolkit.Internal;
 
 internal readonly record struct ElmToolkitServices(
     ServiceProvider ServiceProvider,
