+++ Cql/CoreTests/ExpressionBuilderContextTests.cs
@@ -1,5 +1,5 @@
-using CqlSdkPrototype.Elm.Internal;
 using Hl7.Cql.Abstractions;
+using Hl7.Cql.CodeGeneration.NET.Toolkit.Internal;
 using Hl7.Cql.Compiler;
 using Hl7.Cql.Runtime.Hosting;
 using Hl7.Fhir.Model;
