+++ Cql/CoreTests/LibrarySetExpressionBuilderTests.cs
@@ -1,5 +1,5 @@
-using CqlSdkPrototype.Elm.Internal;
 using Hl7.Cql.Abstractions;
+using Hl7.Cql.CodeGeneration.NET.Toolkit.Internal;
 using Hl7.Cql.Compiler;
 using Hl7.Cql.Primitives;
 using Hl7.Cql.Runtime;
