+++ Cql/CoreTests/PrimitiveTests.cs
@@ -1,5 +1,5 @@
-﻿using CqlSdkPrototype.Elm;
-using Hl7.Cql.Abstractions;
+﻿using Hl7.Cql.Abstractions;
+using Hl7.Cql.CodeGeneration.NET.Toolkit;
 using Hl7.Cql.Compiler;
 using Hl7.Cql.Fhir;
 using Hl7.Cql.Iso8601;
