+++ Cql/CoreTests/QueriesTest.cs
@@ -1,11 +1,11 @@
-﻿using Hl7.Cql.Fhir;
+﻿using Hl7.Cql.CodeGeneration.NET.Toolkit.Internal;
+using Hl7.Cql.Fhir;
 using Hl7.Cql.Primitives;
 using Hl7.Cql.Runtime;
 using Hl7.Cql.ValueSets;
 using Hl7.Fhir.Model;
 using Hl7.Cql.Compiler;
 using Hl7.Cql.Runtime.Hosting;
-using CqlSdkPrototype.Elm.Internal;
 
 namespace CoreTests
 {
