+++ Cql/Cql.CqlToElm/Toolkit/Extensions/CqlToolkitExtensions.Adding.cs
@@ -10,8 +10,17 @@ using Hl7.Cql.Runtime;
 
 namespace Hl7.Cql.CqlToElm.Toolkit.Extensions;
 
+/// <summary>
+/// Provides extension methods for adding CQL libraries to a <see cref="CqlToolkit"/>.
+/// </summary>
 public static partial class CqlToolkitExtensions
 {
+    /// <summary>
+    /// Adds the specified CQL libraries to the <see cref="CqlToolkit"/>.
+    /// </summary>
+    /// <param name="cqlToolkit">The CQL toolkit to add the libraries to.</param>
+    /// <param name="cqlLibraries">The CQL libraries to add.</param>
+    /// <returns>The updated <see cref="CqlToolkit"/>.</returns>
     public static CqlToolkit AddCqlLibraries(
         this CqlToolkit cqlToolkit,
         params CqlLibraryString[] cqlLibraries)
@@ -19,6 +28,14 @@ public static partial class CqlToolkitExtensions
         return cqlToolkit.AddCqlLibraries(cqlLibraries);
     }
 
+    /// <summary>
+    /// Adds CQL libraries from the specified directory to the <see cref="CqlToolkit"/>.
+    /// </summary>
+    /// <param name="cqlToolkit">The CQL toolkit to add the libraries to.</param>
+    /// <param name="directory">The directory to search for CQL files.</param>
+    /// <param name="options">The enumeration options to use when searching for files.</param>
+    /// <param name="filePredicate">An optional predicate to filter the files.</param>
+    /// <returns>The updated <see cref="CqlToolkit"/>.</returns>
     public static CqlToolkit AddCqlLibrariesFromDirectory(
         this CqlToolkit cqlToolkit,
         DirectoryInfo directory,
@@ -30,6 +47,12 @@ public static partial class CqlToolkitExtensions
         return cqlToolkit.AddCqlLibraryFiles(files);
     }
 
+    /// <summary>
+    /// Adds CQL libraries from the specified files to the <see cref="CqlToolkit"/>.
+    /// </summary>
+    /// <param name="cqlToolkit">The CQL toolkit to add the libraries to.</param>
+    /// <param name="files">The files to load the CQL libraries from.</param>
+    /// <returns>The updated <see cref="CqlToolkit"/>.</returns>
     public static CqlToolkit AddCqlLibraryFiles(
         this CqlToolkit cqlToolkit,
         IEnumerable<FileInfo> files)
