+++ Cql/Cql.Packaging/GlobalUsings.cs
@@ -14,10 +14,11 @@ global using ElmLibrarySet = Hl7.Cql.Compiler.LibrarySet;
 global using FhirLibrary = Hl7.Fhir.Model.Library;
 global using FhirMeasure = Hl7.Fhir.Model.Measure;
 global using FhirModelCode = Hl7.Fhir.Model.Code;
+global using FhirModelInfo = Hl7.Fhir.Model.ModelInfo;
 global using FhirResource = Hl7.Fhir.Model.Resource;
 global using Iso8601DateTimePrecision = Hl7.Cql.Iso8601.DateTimePrecision;
 global using SysDateTime = System.DateTime;
 
-global using ElmToolkitConversionDictionary = System.Collections.Immutable.ImmutableDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, CqlSdkPrototype.Elm.ElmToolkitConversionRecord>;
-global using ElmToolkitConversionReadOnlyDictionary = System.Collections.Generic.IReadOnlyDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, CqlSdkPrototype.Elm.ElmToolkitConversionRecord>;
+global using PackagingToolkitConversionsDictionary = System.Collections.Immutable.ImmutableDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, Hl7.Cql.Packaging.Toolkit.PackagingToolkitConversionRecord>;
+global using PackagingToolkitConversionsReadOnlyDictionary = System.Collections.Generic.IReadOnlyDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, Hl7.Cql.Packaging.Toolkit.PackagingToolkitConversionRecord>;
