+++ Cql/Cql.Packaging/Toolkit/Extensions/PackagingToolkitExtensions.Saving.cs
@@ -6,15 +6,21 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
  */
 
-using System.Text.Json;
 using Hl7.Cql.Abstractions.Infrastructure;
-using Hl7.Cql.Packaging;
 using Hl7.Cql.Runtime.IO;
+using Microsoft.Extensions.DependencyInjection;
 
-namespace CqlSdkPrototype.Packaging.Extensions;
+namespace Hl7.Cql.Packaging.Toolkit.Extensions;
 
 public static partial class PackagingToolkitExtensions
 {
+    /// <summary>
+    /// Saves FHIR resources to the specified directory.
+    /// </summary>
+    /// <param name="packagingToolkit">The packaging toolkit instance.</param>
+    /// <param name="directory">The directory where the FHIR resources will be saved.</param>
+    /// <param name="directoryPreparationStrategy">Optional strategy for preparing the directory.</param>
+    /// <returns>The packaging toolkit instance.</returns>
     public static PackagingToolkit SaveFhirResources(
         this PackagingToolkit packagingToolkit,
         DirectoryInfo directory,
@@ -39,6 +45,11 @@ public static partial class PackagingToolkitExtensions
         return packagingToolkit;
     }
 
+    /// <summary>
+    /// Gets the packaging results from the packaging toolkit.
+    /// </summary>
+    /// <param name="packagingToolkit">The packaging toolkit instance.</param>
+    /// <returns>An enumerable of packaging toolkit result records.</returns>
     public static IEnumerable<PackagingToolkitResultRecord> GetPackagingResults(
         this PackagingToolkit packagingToolkit) =>
         packagingToolkit
@@ -46,6 +57,6 @@ public static partial class PackagingToolkitExtensions
             .SelectWhere(kv => kv.Value switch
             {
                 { ResultFhirLibrary: null } => (false, default(PackagingToolkitResultRecord)),
-                var val                     => (true, new(kv.Key, val.ResultFhirLibrary, val.ResultFhirMeasure))
+                var val => (true, new(kv.Key, val.ResultFhirLibrary, val.ResultFhirMeasure))
             });
 }
