+++ Cql/Cql.Packaging/Toolkit/PackagingToolkitConfig.cs
@@ -8,9 +8,13 @@
 
 using Hl7.Cql.Runtime;
 
-namespace CqlSdkPrototype.Packaging;
+namespace Hl7.Cql.Packaging.Toolkit;
 
 #pragma warning disable RS0027
+/// <summary>
+/// Represents the configuration settings for the packaging toolkit.
+/// </summary>
+/// <param name="ErroredEnumerationContinuation">The continuation setting for errored enumerations.</param>
 public record PackagingToolkitConfig(
 #pragma warning restore RS0027
     ErroredEnumerationContinuation ErroredEnumerationContinuation = default)
@@ -20,5 +24,8 @@ public record PackagingToolkitConfig(
     /// </summary>
     public static PackagingToolkitConfig Default { get; } = new PackagingToolkitConfig();
 
+    /// <summary>
+    /// Gets the continuation setting for errored enumerations.
+    /// </summary>
     public ErroredEnumerationContinuation ErroredEnumerationContinuation { get; init; } = ErroredEnumerationContinuation;
 }
