+++ Cql/CqlSdkPrototype/Invocation/Extensions/ElmToolkitExtensions.cs
@@ -6,8 +6,8 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
  */
 
-using CqlSdkPrototype.Elm;
-using CqlSdkPrototype.Elm.Extensions;
+using Hl7.Cql.CodeGeneration.NET.Toolkit;
+using Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 
 namespace CqlSdkPrototype.Invocation.Extensions;
