+++ Cql/CqlToElmTests/(tests)/FunctionDefinitionTest.cs
@@ -1,5 +1,5 @@
-using CqlSdkPrototype.Elm.Extensions;
 using CqlSdkPrototype.Invocation.Extensions;
+using Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 using Hl7.Cql.Elm;
 using Hl7.Cql.Runtime;
