+++ Cql/CqlToElmTests/(tests)/TagsTest.cs
@@ -1,5 +1,5 @@
-using CqlSdkPrototype.Elm.Extensions;
 using CqlSdkPrototype.Invocation.Extensions;
+using Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 using Hl7.Cql.CqlToElm.Toolkit.Extensions;
 
 namespace Hl7.Cql.CqlToElm.Test;
