+++ Cql/CqlToElmTests/Base.cs
@@ -3,11 +3,11 @@ using Hl7.Cql.Compiler;
 using Hl7.Cql.Elm;
 using Hl7.Cql.Fhir;
 using Hl7.Cql.Runtime;
-using CqlSdkPrototype.Elm;
 using Hl7.Cql.Model;
-using CqlSdkPrototype.Elm.Extensions;
 using CqlSdkPrototype.Invocation;
 using CqlSdkPrototype.Invocation.Extensions;
+using Hl7.Cql.CodeGeneration.NET.Toolkit;
+using Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 using Hl7.Cql.CqlToElm.Toolkit;
 
 namespace Hl7.Cql.CqlToElm.Test
