+++ Cql/CqlToElmTests/ElmFluentToolkitExtensions.cs
@@ -1,5 +1,5 @@
-﻿using CqlSdkPrototype.Elm;
-using Hl7.Cql.CodeGeneration.NET;
+﻿using Hl7.Cql.CodeGeneration.NET;
+using Hl7.Cql.CodeGeneration.NET.Toolkit;
 using Hl7.Cql.Compiler;
 using Hl7.Cql.Elm;
 using Hl7.Cql.Runtime;
