+++ Demo/CqlApiExamples/Program.cs
@@ -1,16 +1,16 @@
 ﻿using System.Diagnostics;
 using System.Text;
-using CqlSdkPrototype.Elm;
-using CqlSdkPrototype.Elm.Extensions;
 using CqlSdkPrototype.Invocation;
 using CqlSdkPrototype.Invocation.Extensions;
-using CqlSdkPrototype.Packaging;
-using CqlSdkPrototype.Packaging.Extensions;
 using Hl7.Cql.CodeGeneration.NET;
+using Hl7.Cql.CodeGeneration.NET.Toolkit;
+using Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 using Hl7.Cql.CqlToElm;
 using Hl7.Cql.CqlToElm.Toolkit;
 using Hl7.Cql.CqlToElm.Toolkit.Extensions;
 using Hl7.Cql.Fhir;
+using Hl7.Cql.Packaging.Toolkit;
+using Hl7.Cql.Packaging.Toolkit.Extensions;
 using Hl7.Cql.Runtime;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
